Refactor : 핸들, 유효성 함수 분리
